Address comments (#27513)
