+++ test/tests/auxkernels/functor_material_aux/tests
@@ -4,7 +4,7 @@
     input = real.i
     requirement = "The system shall be able to output a functor to a nodal or elemental variable."
     design = "MaterialRealAux.md"
-    issues = #27513
+    issues = '#27513'
     exodiff = real_out.e
   []
 []
